Chinh sua controller, phan admin dua ve admin, phan manager dua ve manger, chinh sua api & requestMapping
