+++ BE/src/main/java/com/hr_management/hr_management/controller/EmployeeController.java
@@ -31,35 +31,13 @@ public class EmployeeController {
     private final EmployeeService employeeService;
     private final AuthenticationService authenticationService;
 
-    // [ Admin ]
-    // 1. Xem danh sách nhân viên
-    @PreAuthorize("hasRole('ADMIN')")
-    @GetMapping("/api/v1/admin/employees")
-    public ApiResponse<List<EmployeeResponseDTO>> getAll() {
-        return ApiResponse.<List<EmployeeResponseDTO>>builder()
-                .result(employeeService.getAll())
-                .build();
-    }
 
-    // [ MANGER ]
-    // 1. Xem ho so chi tiet nhan vien
-    @PreAuthorize("hasRole('DEPARTMENT_MANAGER')")
-    @GetMapping("/api/v1/manager/employees/{employeeId}")
-    public ApiResponse<EmployeeResponseDTO> getEmployeeByIdByManager(
-            @PathVariable Integer employeeId){
 
-        EmployeeResponseDTO employee = employeeService.getEmployeeById(employeeId);
-
-        return ApiResponse.<EmployeeResponseDTO>builder()
-                .result(employee)
-                .message("Employee retrieved successfully")
-                .build();
-    }
 
     // [ Employee ]
     // 1. Xem hồ sơ cá nhân
 
-    @GetMapping("/api/v1/employee/profile")
+    @GetMapping("/profile")
     public ApiResponse<EmployeeResponseDTO> getMyProfile(JwtAuthenticationToken jwtToken){
         String username = jwtToken.getName();
         EmployeeResponseDTO profile = employeeService.getEmployeeByUserName(username);
@@ -69,7 +47,7 @@ public class EmployeeController {
                 .build();
     }
     // 2. Cập nhật thông tin cá nhân
-    @PutMapping("/api/v1/employee/profile")
+    @PutMapping("/profile")
     public ApiResponse<EmployeeResponseDTO> updateMyProfile(JwtAuthenticationToken jwtToken, @RequestBody UpdateEmployeeProfileRequest request){
         String username = jwtToken.getName();
         EmployeeResponseDTO updatedProfile = employeeService.updateEmployeeProfileByUsername(username, request) ;
@@ -80,12 +58,4 @@ public class EmployeeController {
                 .build();
     }
 
-
-    @PreAuthorize("hasRole('ROLE_DEPARTMENT_MANAGER')")
-    @PostMapping("/employee-department")
-    ApiResponse<GetAllEmployeeDepartmentResponse> getAllEmployeeDepartment(@RequestBody GetAllEmployeeDepartmentRequest request, JwtAuthenticationToken jwtAuthenticationToken){
-        return ApiResponse.<GetAllEmployeeDepartmentResponse>builder()
-                .result(employeeService.getAllEmployeeDepartment(request,jwtAuthenticationToken))
-                .build();
-    }
 }
